backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pillow==9.5.0
opencv-python==4.6.0.66
paddlepaddle==2.5.1
paddleocr==2.6.1.3
google-cloud-vision==3.4.0
numpy==1.24.3
python-dotenv==1.0.0
pydantic==1.10.7
aiofiles==23.1.0
starlette==0.26.1
psycopg2-binary==2.9.6
sqlalchemy==2.0.12
pytest==7.3.1
httpx==0.24.0
